requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
boto3==1.37.1
botocore==1.37.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
dotenv==0.9.9
idna==3.10
jmespath==1.0.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
